update docs, readme
